Add files via upload
Firstly, need to create a GUI interface for the application. For the home page, can create a class named "HomePage" that extends the JFrame class. Also create a class named "SupportPage" that extends the JFrame class for the support page. For the API interface, we can use the HttpClient class in the java.net.http package to send HTTP requests to the API endpoint. We can create a User class to represent the JSON response. For the file-processing layer, we can use the java.nio.file package to read input files from "/app/in" and write output files to "/app/out".